Shuffle init to set environment path inline
