+++ Common/Python/PythonInitializer.cs
@@ -15,7 +15,9 @@
 */
 
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
+using System.IO;
 using Python.Runtime;
 
 namespace QuantConnect.Python
@@ -26,24 +28,44 @@ namespace QuantConnect.Python
     public static class PythonInitializer
     {
         // Used to allow multiple Python unit and regression tests to be run in the same test run
-        private static bool _isBeginAllowThreadsCalled;
+        private static bool _initialized;
 
         /// <summary>
         /// Initialize the Python.NET library
         /// </summary>
-        public static void Initialize()
+        public static void Initialize(string assemblyPath = "")
         {
-            if (!_isBeginAllowThreadsCalled)
+            if (!_initialized)
             {
+                if (assemblyPath != string.Empty)
+                {
+                    // Set the python path for loading python algorithms.
+                    var pythonPath = new List<string>
+                    {
+                        new FileInfo(assemblyPath).Directory.FullName,
+                        new DirectoryInfo(Environment.CurrentDirectory).FullName,
+                    };
+
+                    // Don't include an empty environment variable in pythonPath, otherwise the PYTHONPATH
+                    // environment variable won't be used in the module import process
+                    var pythonPathEnvironmentVariable = Environment.GetEnvironmentVariable("PYTHONPATH");
+                    if (!string.IsNullOrEmpty(pythonPathEnvironmentVariable))
+                    {
+                        pythonPath.Add(pythonPathEnvironmentVariable);
+                    }
+
+                    Environment.SetEnvironmentVariable("PYTHONPATH", string.Join(OS.IsLinux ? ":" : ";", pythonPath));
+                }
+
                 var benchmark = Stopwatch.StartNew();
                 PythonEngine.Initialize();
+                Logging.Log.Trace("PythonInitializer.Initialize(): Python Engine Initialized in " + benchmark.Elapsed.TotalSeconds + "s. Thread: " + System.Threading.Thread.CurrentThread.Name);
                 benchmark.Restart();
-                Logging.Log.Trace("PythonInitializer(): Python Engine Initialized in " + benchmark.Elapsed.TotalSeconds + "s. Thread: " + System.Threading.Thread.CurrentThread.Name);
 
                 // required for multi-threading usage
                 PythonEngine.BeginAllowThreads();
-                Logging.Log.Trace("PythonInitializer():  Allow threading completed in " + benchmark.Elapsed.TotalSeconds + "s. Thread: " + System.Threading.Thread.CurrentThread.Name);
-                _isBeginAllowThreadsCalled = true;
+                Logging.Log.Trace("PythonInitializer.Initialize():  Allow threading completed in " + benchmark.Elapsed.TotalSeconds + "s. Thread: " + System.Threading.Thread.CurrentThread.Name);
+                _initialized = true;
             }
         }
     }
